Fix duplicate key error in snippets
Atom throws an error as soon as it loads this package, because it won't load duplicate keys in a snippets file. When you want to define more than one prefix to a given snippet, it's a known issue/request (https://github.com/atom/snippets/issues/107).

In the meantime, the workaround is to duplicate as was done here, but name the snippets differently.